// dev/@types/ActionTestSummaryGroup.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface ActionTestSummaryGroup {
  name?: string
  identifier?: string
  duration: number
  subtests: ActionTestSummaryIdentifiableObject[]
}
export interface ActionTestSummaryIdentifiableObject {
  name?: string
  identifier?: string
}